refactor(tests): Create `UploadedFilesPsr7Test` class to organize and enhance tests for uploaded files handling in PSR-7 requests.
