+++ tests/adapter/UploadedFilesPsr7Test.php
@@ -1,0 +1,407 @@
+<?php
+
+declare(strict_types=1);
+
+namespace yii2\extensions\psrbridge\tests\adapter;
+
+use PHPUnit\Framework\Attributes\Group;
+use yii\web\UploadedFile;
+use yii2\extensions\psrbridge\http\Request;
+use yii2\extensions\psrbridge\tests\support\FactoryHelper;
+use yii2\extensions\psrbridge\tests\TestCase;
+
+#[Group('adapter')]
+#[Group('uploaded-files')]
+final class UploadedFilesPsr7Test extends TestCase
+{
+    public function testReturnMultipleUploadedFilesWithDifferentStructures(): void
+    {
+        $tmpFile1 = $this->createTmpFile();
+
+        $tmpPathFile1 = stream_get_meta_data($tmpFile1)['uri'];
+        $tmpFileSize1 = filesize($tmpPathFile1);
+
+        $tmpFile2 = $this->createTmpFile();
+
+        $tmpPathFile2 = stream_get_meta_data($tmpFile2)['uri'];
+        $tmpFileSize2 = filesize($tmpPathFile2);
+
+        self::assertNotFalse(
+            $tmpFileSize1,
+            "'filesize' for 'test1.txt' should not be 'false'.",
+        );
+        self::assertNotFalse(
+            $tmpFileSize2,
+            "'filesize' for 'test2.php' should not be 'false'.",
+        );
+
+        $uploadedFiles = [
+            'simple1' => FactoryHelper::createUploadedFile(
+                'simple1.txt',
+                'text/plain',
+                $tmpPathFile1,
+                size: $tmpFileSize1,
+            ),
+            'simple2' => FactoryHelper::createUploadedFile(
+                'simple2.php',
+                'application/x-php',
+                $tmpPathFile2,
+                size: $tmpFileSize2,
+            ),
+            'nested' => [
+                'level1' => FactoryHelper::createUploadedFile(
+                    'nested1.txt',
+                    'text/plain',
+                    $tmpPathFile1,
+                    size: $tmpFileSize1,
+                ),
+                'level2' => FactoryHelper::createUploadedFile(
+                    'nested2.php',
+                    'application/x-php',
+                    $tmpPathFile2,
+                    size: $tmpFileSize2,
+                ),
+            ],
+            'array_files' => [
+                FactoryHelper::createUploadedFile(
+                    'array1.txt',
+                    'text/plain',
+                    $tmpPathFile1,
+                    size: $tmpFileSize1,
+                ),
+                FactoryHelper::createUploadedFile(
+                    'array2.php',
+                    'application/x-php',
+                    $tmpPathFile2,
+                    size: $tmpFileSize2,
+                ),
+            ],
+        ];
+
+        $request = new Request();
+
+        $request->setPsr7Request(
+            FactoryHelper::createRequest('POST', '/upload')->withUploadedFiles($uploadedFiles),
+        );
+
+        $convertedFiles = $request->getUploadedFiles();
+
+        self::assertCount(
+            4,
+            $convertedFiles,
+            "Should return all '4' top-level items in the 'UploadedFiles' array, matching the original structure.",
+        );
+        self::assertInstanceOf(
+            UploadedFile::class,
+            $convertedFiles['simple1'] ?? null,
+            "'simple1' should be an instance of 'UploadedFile', representing a single uploaded file.",
+        );
+        self::assertInstanceOf(
+            UploadedFile::class,
+            $convertedFiles['simple2'] ?? null,
+            "'simple2' should be an instance of 'UploadedFile', representing a single uploaded file.",
+        );
+        self::assertIsArray(
+            $convertedFiles['nested'] ?? null,
+            "'nested' should be an array, representing a nested structure of uploaded files.",
+        );
+        self::assertCount(
+            2,
+            $convertedFiles['nested'],
+            "'nested' array should contain exactly '2' items: 'level1' and 'level2'.",
+        );
+        self::assertInstanceOf(
+            UploadedFile::class,
+            $convertedFiles['nested']['level1'] ?? null,
+            "'nested['level1']' should be an instance of 'UploadedFile', representing a nested uploaded file.",
+        );
+        self::assertInstanceOf(
+            UploadedFile::class,
+            $convertedFiles['nested']['level2'] ?? null,
+            "'nested['level2']' should be an instance of 'UploadedFile', representing a nested uploaded file.",
+        );
+        self::assertIsArray(
+            $convertedFiles['array_files'] ?? null,
+            "'array_files' should be an array, representing a list of uploaded files.",
+        );
+        self::assertCount(
+            2,
+            $convertedFiles['array_files'],
+            "'array_files' should contain exactly '2' items, each representing an uploaded file.",
+        );
+        self::assertInstanceOf(
+            UploadedFile::class,
+            $convertedFiles['array_files'][0] ?? null,
+            "'array_files[0]' should be an instance of 'UploadedFile', representing the first file in the array.",
+        );
+        self::assertInstanceOf(
+            UploadedFile::class,
+            $convertedFiles['array_files'][1] ?? null,
+            "'array_files[1]' should be an instance of 'UploadedFile', representing the second file in the array.",
+        );
+    }
+
+    public function testReturnUploadedFilesRecursivelyConvertsNestedArrays(): void
+    {
+        $file1 = dirname(__DIR__) . '/support/stub/files/test1.txt';
+        $file2 = dirname(__DIR__) . '/support/stub/files/test2.php';
+        $size1 = filesize($file1);
+        $size2 = filesize($file2);
+
+        self::assertNotFalse(
+            $size1,
+            "'filesize' for 'test1.txt' should not be 'false'.",
+        );
+        self::assertNotFalse(
+            $size2,
+            "'filesize' for 'test2.php' should not be 'false'.",
+        );
+
+        $uploadedFile1 = FactoryHelper::createUploadedFile('test1.txt', 'text/plain', $file1, size: $size1);
+        $uploadedFile2 = FactoryHelper::createUploadedFile('test2.php', 'application/x-php', $file2, size: $size2);
+
+        $deepNestedFiles = [
+            'docs' => [
+                'sub' => [
+                    'file1' => $uploadedFile1,
+                    'file2' => $uploadedFile2,
+                ],
+            ],
+        ];
+
+        $psr7Request = FactoryHelper::createRequest('POST', '/upload')->withUploadedFiles($deepNestedFiles);
+
+        $request = new Request();
+
+        $request->setPsr7Request($psr7Request);
+
+        $deepNestedUploadedFiles = $request->getUploadedFiles();
+
+        $expectedUploadedFiles = [
+            'file1' => [
+                'name' => 'test1.txt',
+                'type' => 'text/plain',
+                'tempName' => $file1,
+                'error' => UPLOAD_ERR_OK,
+                'size' => $size1,
+            ],
+            'file2' => [
+                'name' => 'test2.php',
+                'type' => 'application/x-php',
+                'tempName' => $file2,
+                'error' => UPLOAD_ERR_OK,
+                'size' => $size2,
+            ],
+        ];
+
+        $runtimePath = dirname(__DIR__, 2) . '/runtime';
+
+        foreach ($deepNestedUploadedFiles as $nestedUploadFiles) {
+            if (is_array($nestedUploadFiles)) {
+                foreach ($nestedUploadFiles as $uploadedFiles) {
+                    if (is_array($uploadedFiles)) {
+                        foreach ($uploadedFiles as $name => $uploadedFile) {
+                            self::assertInstanceOf(
+                                UploadedFile::class,
+                                $uploadedFile,
+                                "Uploaded file '{$name}' should be an instance of '" . UploadedFile::class . "'.",
+                            );
+                            self::assertSame(
+                                $expectedUploadedFiles[$name]['name'] ?? null,
+                                $uploadedFile->name,
+                                "Uploaded file '{$name}' should have the expected client filename.",
+                            );
+                            self::assertSame(
+                                $expectedUploadedFiles[$name]['type'] ?? null,
+                                $uploadedFile->type,
+                                "Uploaded file '{$name}' should have the expected client media type.",
+                            );
+                            self::assertSame(
+                                $expectedUploadedFiles[$name]['tempName'] ?? null,
+                                $uploadedFile->tempName,
+                                "Uploaded file '{$name}' should have the expected temporary name.",
+                            );
+                            self::assertSame(
+                                $expectedUploadedFiles[$name]['error'] ?? null,
+                                $uploadedFile->error,
+                                "Uploaded file '{$name}' should have the expected error code.",
+                            );
+                            self::assertSame(
+                                $expectedUploadedFiles[$name]['size'] ?? null,
+                                $uploadedFile->size,
+                                "Uploaded file '{$name}' should have the expected size.",
+                            );
+                            self::assertTrue(
+                                $uploadedFile->saveAs("{$runtimePath}/{$uploadedFile->name}", false),
+                                "Uploaded file '{$uploadedFile->name}' should be saved to the runtime directory " .
+                                'successfully.',
+                            );
+                            self::assertFileExists(
+                                "{$runtimePath}/{$uploadedFile->name}",
+                                "Uploaded file '{$uploadedFile->name}' should exist in the runtime directory after " .
+                                'saving.',
+                            );
+                        }
+                    }
+                }
+            }
+        }
+    }
+    public function testReturnUploadedFilesWhenAdapterIsSet(): void
+    {
+        $file1 = dirname(__DIR__) . '/support/stub/files/test1.txt';
+        $file2 = dirname(__DIR__) . '/support/stub/files/test2.php';
+        $size1 = filesize($file1);
+        $size2 = filesize($file2);
+
+        self::assertNotFalse(
+            $size1,
+            "'filesize' for 'test1.txt' should not be 'false'.",
+        );
+        self::assertNotFalse(
+            $size2,
+            "'filesize' for 'test2.php' should not be 'false'.",
+        );
+
+        $uploadedFile1 = FactoryHelper::createUploadedFile('test1.txt', 'text/plain', $file1, size: $size1);
+        $uploadedFile2 = FactoryHelper::createUploadedFile('test2.php', 'application/x-php', $file2, size: $size2);
+        $psr7Request = FactoryHelper::createRequest('POST', '/upload');
+
+        $psr7Request = $psr7Request->withUploadedFiles(
+            [
+                'file1' => $uploadedFile1,
+                'file2' => $uploadedFile2,
+            ],
+        );
+
+        $request = new Request();
+
+        $request->setPsr7Request($psr7Request);
+
+        $uploadedFiles = $request->getUploadedFiles();
+
+        $expectedNames = [
+            'file1',
+            'file2',
+        ];
+        $expectedUploadedFiles = [
+            'file1' => [
+                'name' => 'test1.txt',
+                'type' => 'text/plain',
+                'tempName' => $file1,
+                'error' => UPLOAD_ERR_OK,
+                'size' => $size1,
+            ],
+            'file2' => [
+                'name' => 'test2.php',
+                'type' => 'application/x-php',
+                'tempName' => $file2,
+                'error' => UPLOAD_ERR_OK,
+                'size' => $size2,
+            ],
+        ];
+
+        $runtimePath = dirname(__DIR__, 2) . '/runtime';
+
+        foreach ($uploadedFiles as $name => $uploadedFile) {
+            self::assertContains(
+                $name,
+                $expectedNames,
+                "Uploaded file name '{$name}' should be in the expected names list.",
+            );
+            self::assertInstanceOf(
+                UploadedFile::class,
+                $uploadedFile,
+                "Uploaded file '{$name}' should be an instance of '" . UploadedFile::class . "'.",
+            );
+            self::assertSame(
+                $expectedUploadedFiles[$name]['name'] ?? null,
+                $uploadedFile->name,
+                "Uploaded file '{$name}' should have the expected client filename.",
+            );
+            self::assertSame(
+                $expectedUploadedFiles[$name]['type'] ?? null,
+                $uploadedFile->type,
+                "Uploaded file '{$name}' should have the expected client media type.",
+            );
+            self::assertSame(
+                $expectedUploadedFiles[$name]['tempName'] ?? null,
+                $uploadedFile->tempName,
+                "Uploaded file '{$name}' should have the expected temporary name.",
+            );
+            self::assertSame(
+                $expectedUploadedFiles[$name]['error'] ?? null,
+                $uploadedFile->error,
+                "Uploaded file '{$name}' should have the expected error code.",
+            );
+            self::assertSame(
+                $expectedUploadedFiles[$name]['size'] ?? null,
+                $uploadedFile->size,
+                "Uploaded file '{$name}' should have the expected size.",
+            );
+            self::assertTrue(
+                $uploadedFile->saveAs("{$runtimePath}/{$uploadedFile->name}", false),
+                "Uploaded file '{$uploadedFile->name}' should be saved to the runtime directory successfully.",
+            );
+            self::assertFileExists(
+                "{$runtimePath}/{$uploadedFile->name}",
+                "Uploaded file '{$uploadedFile->name}' should exist in the runtime directory after saving.",
+            );
+        }
+    }
+
+    public function testReturnUploadedFileWithZeroSizeWhenPsr7FileSizeIsNull(): void
+    {
+        $file1 = dirname(__DIR__) . '/support/stub/files/test1.txt';
+
+        $uploadedFile1 = FactoryHelper::createUploadedFile('test1.txt', 'text/plain', $file1);
+        $psr7Request = FactoryHelper::createRequest('POST', '/upload');
+
+        $psr7Request = $psr7Request->withUploadedFiles(['test_file' => $uploadedFile1]);
+
+        $request = new Request();
+
+        $request->setPsr7Request($psr7Request);
+
+        $uploadedFiles = $request->getUploadedFiles();
+
+        self::assertArrayHasKey(
+            'test_file',
+            $uploadedFiles,
+            "Uploaded files array should contain the 'test_file' key.",
+        );
+
+        $uploadedFile = $uploadedFiles['test_file'] ?? null;
+
+        self::assertInstanceOf(
+            UploadedFile::class,
+            $uploadedFile,
+            "Value for 'test_file' should be an instance of 'UploadedFile'.",
+        );
+        self::assertSame(
+            UPLOAD_ERR_OK,
+            $uploadedFile->error,
+            "'UploadedFile' 'error' property should be 'UPLOAD_ERR_OK'.",
+        );
+        self::assertSame(
+            'test1.txt',
+            $uploadedFile->name,
+            "'UploadedFile' 'name' property should match the original filename.",
+        );
+        self::assertSame(
+            0,
+            $uploadedFile->size,
+            "'UploadedFile' 'size' should default to 0 when PSR-7 file 'getSize()' returns 'null'.",
+        );
+        self::assertSame(
+            $file1,
+            $uploadedFile->tempName,
+            "'UploadedFile' 'tempName' should match the original file path.",
+        );
+        self::assertSame(
+            'text/plain',
+            $uploadedFile->type,
+            "'UploadedFile' 'type' should match the original MIME type.",
+        );
+    }
+}
